fix: resolve remaining #468 review feedback
