+++ src/types/index.ts
@@ -1,5 +1,11 @@
-/** Core data types for the kbexplorer knowledge graph. */
+/**
+ * Core data types for the kbexplorer knowledge graph.
+ *
+ * The graph contract stays engine-free in this template so representation and
+ * serialization layers can consume it without pulling in engine runtime logic.
+ */
 
+import type { NodeLayer } from './node-layer';
 import {
   buildJsonLd,
   type AccessClassification,
@@ -44,7 +50,9 @@ import {
   type PresentationTokens,
   CANVAS_TARGET,
 } from '@anokye-labs/kbexplorer-core';
-export type NodeLayer = 'file' | 'content' | 'work';
+
+export type { NodeLayer } from './node-layer';
+export type { KnownRelation } from './known-relation';
 
 export interface KBNode extends CoreKBNode {
  layer?: NodeLayer;
@@ -58,10 +66,9 @@ export interface KBGraph extends CoreKBGraph {
 * Re-export the pure graph + config contract from `@anokye-labs/kbexplorer-core`
  * so existing `../types` imports keep working unchanged. The default-config
  * logic and the pure graph projections below (collapse/trim) stay template-local
- * but engine-free; styling, layer and view representation now live under
- * `../representation` (Phase 2 / F2 #309) so this module imports nothing from the
- * engine at load.
- */
+* but engine-free; styling, layer, and view representation now live under
+* `../representation` so this module stays free of engine runtime imports.
+*/
 export {
   buildJsonLd,
   type AccessClassification,
@@ -151,29 +158,6 @@ export interface NodeMapEntry {
 export interface NodeMap {
   nodes: NodeMapEntry[];
 }
-
-/**
- * The open relationship taxonomy carried by {@link KBEdge.relation}.
- *
- * These six relations come from the content model and are rendered in the
- * legend data-drivenly. `relation` is an open string — unknown relations still
- * render with a sensible default style.
- */
-export type KnownRelation =
-  | 'leads'
-  | 'staffs'
-  | 'reports-to'
-  | 'structural'
-  | 'derived'
-  | 'deprecated'
-  // Work-graph organizational-layer relations (#233)
-  | 'owns'
-  | 'has-priority'
-  | 'tracked-in'
-  // Person-node active-work relations (#235)
-  | 'assigned-to'
-  | 'authored'
-  | 'member-of';
 
 /**
  * Collapse specified clusters into single summary nodes.
