+++ src/types/node-layer.ts
@@ -1,0 +1,1 @@
+export type NodeLayer = 'file' | 'content' | 'work';
